+++ scripts/aerospike-client-c.sh
@@ -1,6 +1,6 @@
-#! /bin/bash
+#!/bin/bash
 ################################################################################
-# Copyright 2013-2014 Aerospike, Inc.
+# Copyright 2013-2016 Aerospike, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -25,13 +25,16 @@ AEROSPIKE_C_VERSION=${AEROSPIKE_C_VERSION:-'4.0.7'}
 ################################################################################
 
 CWD=$(pwd)
+SCRIPT_DIR=$(dirname $0)
 AEROSPIKE=${CWD}/aerospike-client-c
-
-unset PKG_DIST
-unset PKG_TYPE
-unset PKG_PATH
-
+LIB_PATH=${PREFIX}
 LUA_PATH=${AEROSPIKE_LUA_PATH}
+
+DOWNLOAD=${DOWNLOAD:=0}
+COPY_FILES=1
+
+unset PKG_TYPE
+
 
 ################################################################################
 #
@@ -39,190 +42,69 @@ LUA_PATH=${AEROSPIKE_LUA_PATH}
 #
 ################################################################################
 
-detect_linux()
-{
-  # check to see if `lsb_release` is available.
-  if [ ! -z "$(which lsb_release)" ]; then
-
-    # We have LSB, so use it.
-    DIST_IDEN=$(lsb_release -is | tr '[:upper:]' '[:lower:]')
-    DIST_VERS=$(lsb_release -rs | cut -d. -f1 )
-    DIST_NAME=${DIST_IDEN}${DIST_VERS}
-
-    case ${DIST_NAME} in
-
-      "centos6" | "redhatenterpriseserver6" | "centos7" | "fedora20" | "fedora21" | "fedora22" | "fedora23" | "korora22" | "korora23")
-        echo "el6" "rpm"
-        return 0
-        ;;
-
-      "debian6" )
-        echo "debian6" "deb"
-        return 0
-        ;;
-
-
-      "debian7" | "debian8" )
-        echo "debian7" "deb"
-        return 0
-        ;;
-
-      "ubuntu12" | "ubuntu13" | "ubuntu14" | "ubuntu15" | "ubuntu16" | "linuxmint17" )
-        echo "ubuntu12"  "deb"
-        return 0
-        ;;
-
-      * )
-        echo "error: ${DIST_NAME} is not supported."
-        return 1
-        ;;
-
-    esac
-  fi
-
-  # no LSB, check for /etc/redhat-release
-  if [ -f /etc/redhat-release ]; then
-    dist=$(cat /etc/redhat-release | tr '[:upper:]' '[:lower:]')
-    case ${dist} in
-
-      "centos"* | "red hat enterprise linux"* | "fedora"* | "korora"*)
-	echo "el6" "rpm"
-	return 0
-	;;
-
-       * )
-	echo "error: ${dist} is not supported."
-	return 1
-	;;
-     esac
-   fi
-
-
-  # Ok, no LSB, so check for /etc/issue
-  if [ -f /etc/issue ]; then
-    dist=$(cat /etc/issue | tr '[:upper:]' '[:lower:]')
-    case ${dist} in
-
-      "centos"* | "red hat enterprise linux"* | "fedora"* | "korora"*)
-        echo "el6" "rpm"
-        return 0
-        ;;
-
-      "debian"* )
-        vers=$(cat /etc/debian_version)
-        case ${vers} in
-          "7."* | "8."* )
-            echo "debian7" "deb"
-            return 0
-            ;;
-
-          * )
-            echo "error: Debian ${vers} is not supported."
-            return 1
-            ;;
-        esac
-        ;;
-
-      "ubuntu"* )
-        vers=$(lsb_release -r -s)
-        case ${vers} in
-          "12."* | "13."* | "14.*" )
-            echo "ubuntu12"  "deb"
-            return 0
-            ;;
-
-          * )
-            echo "error: Ubuntu ${vers} is not supported."
-            return 1
-            ;;
-        esac
-        ;;
-
-      "amazon linux"* )
-        echo "el6" "rpm"
-        return 0
-        ;;
-
-      * )
-        echo "error: ${DIST_NAME} is not supported."
-        return 1
-        ;;
-
-    esac
-  fi
-
-  echo "error: Linux Distro not supported"
-  return 1
+has_cmd() {
+  hash "$1" 2> /dev/null
 }
 
-download()
-{
-  # create the package directory
+download() {
   mkdir -p ${AEROSPIKE}/package
 
-  # check to see if `curl` or `wget` is available.
-  curl_path=`which curl`
-  has_curl=$?
+  URL="http://artifacts.aerospike.com/aerospike-client-c/${AEROSPIKE_C_VERSION}/${PKG_ARTIFACT}"
 
-  # check for `wget`
-  wget_path=`which wget`
-  has_wget=$?
+  printf "info: downloading '%s' to '%s'\n" "${URL}" "${AEROSPIKE}/package/${PKG_ARTIFACT}"
 
-  # Check
-  if [ $has_curl != 0 ] && [ $has_wget != 0 ]; then
-    echo "error: Not able to find 'curl' or `wget`. Either is required to download the package."
+  if has_cmd curl; then
+    curl -L ${URL} > ${AEROSPIKE}/package/${PKG_ARTIFACT}
+    if [ $? != 0 ]; then
+      echo "error: Unable to download package from '${URL}'"
+      exit 1
+    fi
+  elif has_cmd wget; then
+    wget -O ${AEROSPIKE}/package/${PKG_ARTIFACT} ${URL}
+    if [ $? != 0 ]; then
+      echo "error: Unable to download package from '${URL}'"
+      exit 1
+    fi
+  else
+    echo "error: Not able to find 'curl' or 'wget'. Either is required to download the package."
     exit 1
-  fi
-
-  # Compose the URL for the client tgz
-  URL="http://www.aerospike.com/download/client/c/${AEROSPIKE_C_VERSION}/artifact/${PKG_DIST}"
-
-  # Download and extract the client tgz.
-  # Use non-slient mode to show progress about the download. Important for slower networks.
-  printf "info: downloading '${URL}' to '${AEROSPIKE}/package/aerospike-client-c.tgz'\n"
-
-  if [ $has_curl == 0 ]; then
-    curl -L ${URL} > ${AEROSPIKE}/package/aerospike-client-c.tgz
-    if [ $? != 0 ]; then
-      echo "error: Unable to download package from '${URL}'"
-      exit 1
-    fi
-  elif [ $has_wget == 0 ]; then
-    wget -O ${AEROSPIKE}/package/aerospike-client-c.tgz ${URL}
-    if [ $? != 0 ]; then
-      echo "error: Unable to download package from '${URL}'"
-      exit 1
-    fi
   fi
 
   return 0
 }
 
+function check_lib_path() {
+  [ -d "$1" ] && [ -f "$1/lib/libaerospike.a" ] && [ -f "$1/include/aerospike/aerospike.h" ]
+}
+
+function check_lua_path() {
+  [ -d "$1" ] && [ -f "$1/aerospike.lua" ]
+}
+
 ################################################################################
-# PREFIX is not defined, so we want to see if we can derive it.
+# LIB_PATH is not defined, so we want to see if we can derive it.
 ################################################################################
 
-if [ ! $DOWNLOAD ] && [ ! $PREFIX ]; then
-  if [ -d ${AEROSPIKE} ] && [ -f ${AEROSPIKE}/package/usr/lib/libaerospike.a ] && [ -f ${AEROSPIKE}/package/usr/include/aerospike/aerospike.h ]; then
-    # first, check to see if there is a local client
-    PREFIX=${AEROSPIKE}/package/usr
-    if [ -f ${AEROSPIKE}/package/opt/aerospike/client/sys/udf/lua/aerospike.lua ]; then
-      LUA_PATH=${AEROSPIKE}/package/opt/aerospike/client/sys/udf/lua
-    elif [ -f ${AEROSPIKE}/package/usr/local/aerospike/client/sys/udf/lua/aerospike.lua ]; then
-      LUA_PATH=${AEROSPIKE}/package/usr/local/aerospike/client/sys/udf/lua
+if [ $DOWNLOAD == 0 ] && [ -z $LIB_PATH ]; then
+  # first, check to see if there is a local client
+  if check_lib_path ${AEROSPIKE}; then
+    LIB_PATH=${AEROSPIKE}
+    if check_lua_path ${AEROSPIKE}/lua; then
+      LUA_PATH=${AEROSPIKE}/lua
     fi
-  elif [ -f /usr/lib/libaerospike.a ] && [ -f /usr/include/aerospike/aerospike.h ]; then
-    # next, check to see if there is an installed client
-    PREFIX=/usr
-    if [ -f /opt/aerospike/client/sys/udf/lua/aerospike.lua ]; then
+    COPY_FILES=0
+  # next, check to see if there is an installed client
+  elif check_lib_path "/user"; then
+    LIB_PATH=/usr
+    if check_lua_path /opt/aerospike/client/sys/udf/lua; then
       LUA_PATH=/opt/aerospike/client/sys/udf/lua
-    elif [ -f /usr/local/aerospike/client/sys/udf/lua/aerospike.lua ]; then
+    elif check_lua_path /usr/local/aerospike/client/sys/udf/lua; then
       LUA_PATH=/usr/local/aerospike/client/sys/udf/lua
     fi
   fi
 
   # If we can't find it, then download it.
-  if [ ! $PREFIX ]; then
+  if [ ! $LIB_PATH ]; then
     DOWNLOAD=1
   fi
 fi
@@ -241,14 +123,30 @@ if [ $DOWNLOAD ] && [ $DOWNLOAD == 1 ]; then
     # LINUX
     ############################################################################
     "linux" )
-      result=$(detect_linux)
+      PKG_DIST=$($SCRIPT_DIR/os_version)
       if [ $? -ne 0 ]; then
-        printf "$result\n" >&2
+        printf "%s\n" "$PKG_DIST" >&2
         exit 1
       fi
 
-      IFS=" " read PKG_DIST PKG_TYPE <<< "${result}"
-      PKG_PATH=${AEROSPIKE}/package/usr
+      case $PKG_DIST in
+        "el"* )
+          RPM_VERSION="${AEROSPIKE_C_VERSION//-/_}-1"
+          PKG_ARTIFACT="aerospike-client-c-libuv-devel-${RPM_VERSION}.${PKG_DIST}.x86_64.rpm"
+          PKG_TYPE="rpm"
+          ;;
+        "debian"* )
+          PKG_ARTIFACT="aerospike-client-c-libuv-devel-${AEROSPIKE_C_VERSION}.${PKG_DIST}.x86_64.deb"
+          PKG_TYPE="deb"
+          ;;
+        "ubuntu"* )
+          OS_VERSION_LONG=$($SCRIPT_DIR/os_version -long)
+          PKG_ARTIFACT="aerospike-client-c-libuv-devel-${AEROSPIKE_C_VERSION}.${OS_VERSION_LONG}.x86_64.deb"
+          PKG_TYPE="deb"
+          ;;
+      esac
+
+      LIB_PATH=${AEROSPIKE}/package/usr
       LUA_PATH=${AEROSPIKE}/package/opt/aerospike/client/sys/udf/lua
       ;;
 
@@ -256,9 +154,9 @@ if [ $DOWNLOAD ] && [ $DOWNLOAD == 1 ]; then
     # MAC OS X
     ############################################################################
     "darwin" )
-      PKG_DIST="mac"
+      PKG_ARTIFACT="aerospike-client-c-libuv-devel-${AEROSPIKE_C_VERSION}.pkg"
       PKG_TYPE="pkg"
-      PKG_PATH=${AEROSPIKE}/package/usr/local
+      LIB_PATH=${AEROSPIKE}/package/usr/local
       LUA_PATH=${AEROSPIKE}/package/usr/local/aerospike/client/sys/udf/lua
       ;;
 
@@ -277,65 +175,48 @@ if [ $DOWNLOAD ] && [ $DOWNLOAD == 1 ]; then
   # We will then move the files to the correct location for building.
   ##############################################################################
 
-  if [ -d ${AEROSPIKE}/package/usr ]; then
-    printf "warning: 'aerospike-client-c/package/usr' directory exists.\n"
+  if check_lib_path ${AEROSPIKE}; then
+    printf "warning: '%s' directory exists.\n" "${AEROSPIKE}"
     printf "warning: \n"
     printf "warning: We will be using this directory, rather than downloading a\n"
     printf "warning: new package. If you would like to download a new package\n"
-    printf "warning: then please remove the 'aerospike-client' directory and any \n"
-    printf "warning: 'aerospike-client.tgz' file in this directory.\n"
-    printf "warning: \n"
+    printf "warning: please remove the '%s' directory.\n" $(basename ${AEROSPIKE})
   else
 
     ##############################################################################
     # DOWNLOAD TGZ
     ##############################################################################
 
-    if [ -f ${AEROSPIKE}/package/aerospike-client-c.tgz ]; then
-      printf "warning: 'aerospike-client-c/package/aerospike-client-c.tgz' file exists.\n"
+    if [ -f ${AEROSPIKE}/package/${PKG_ARTIFACT} ]; then
+      printf "warning: '%s' file exists.\n" "${AEROSPIKE}/package/${PKG_ARTIFACT}"
       printf "warning: \n"
       printf "warning: We will be using this package, rather than downloading a new package.\n"
-      printf "warning: If you would like to download a new package, then please remove.\n"
-      printf "warning: 'aerospike-client.tgz' from this directory.\n"
-      printf "warning: \n"
+      printf "warning: If you would like to download a new package please remove\n"
+      printf "warning: the package file.\n"
     else
       download
-
     fi
-
-    ##############################################################################
-    # EXTRACT DEVEL INSTALLER
-    ##############################################################################
-
-    # let's go into the directory
-    cd ${AEROSPIKE}/package
-
-    # Find the `devel` installer package in the client tgz
-    INST_PATH=$(tar -tzf aerospike-client-c.tgz | grep -e ".*libuv-devel-.*\.${PKG_TYPE}")
-
-    # Extract the `devel` installer package from the client tgz
-    printf "info: extracting '${INST_PATH}' from 'aerospike-client-c.tgz'\n"
-    tar -xzf aerospike-client-c.tgz --strip=1 ${INST_PATH}
 
     ##############################################################################
     # EXTRACT FILES FROM DEVEL INSTALLER
     ##############################################################################
 
+    # let's go into the directory
+    cd ${AEROSPIKE}/package
+
     # Extract the contents of the `devel` installer package into `aerospike-client`
+    printf "info: extracting files from '%s'\n" ${PKG_ARTIFACT}
     case ${PKG_TYPE} in
       "rpm" )
-        printf "info: extracting files from '${INST_PATH}'\n"
-        rpm2cpio aerospike-client-c-libuv-devel-*.rpm | cpio -idm --no-absolute-filenames
+        rpm2cpio ${PKG_ARTIFACT} | cpio -idmu --no-absolute-filenames --quiet
         ;;
       "deb" )
-        printf "info: extracting files from '${INST_PATH}'\n"
-        dpkg -x aerospike-client-c-libuv-devel-*.deb .
+        dpkg -x ${PKG_ARTIFACT} .
         ;;
       "pkg" )
-        printf "info: extracting files from '${INST_PATH}'\n"
-        xar -xf aerospike-client-c-libuv-devel-*.pkg
-        cat Payload | gunzip -dc | cpio -i
-        rm Bom PackageInfo Payload
+        xar -xf ${PKG_ARTIFACT} Payload
+        cpio -idmu -I Payload --quiet
+        rm Payload
         ;;
     esac
 
@@ -344,40 +225,37 @@ if [ $DOWNLOAD ] && [ $DOWNLOAD == 1 ]; then
 
   fi
 
-  PREFIX=${PKG_PATH}
-
 fi
 
 ################################################################################
 # PERFORM CHECKS
 ################################################################################
 
-AEROSPIKE_LIBRARY=${PREFIX}/lib/libaerospike.a
-AEROSPIKE_INCLUDE=${PREFIX}/include/aerospike
+AEROSPIKE_LIBRARY=${LIB_PATH}/lib/libaerospike.a
+AEROSPIKE_INCLUDE=${LIB_PATH}/include
 AEROSPIKE_LUA=${LUA_PATH}
 
 printf "\n" >&1
-
 printf "CHECK\n" >&1
 
 if [ -f ${AEROSPIKE_LIBRARY} ]; then
-  printf "   [✓] ${AEROSPIKE_LIBRARY}\n" >&1
+  printf "   [✓] %s\n" "${AEROSPIKE_LIBRARY}" >&1
 else
-  printf "   [✗] ${AEROSPIKE_LIBRARY}\n" >&1
+  printf "   [✗] %s\n" "${AEROSPIKE_LIBRARY}" >&1
   FAILED=1
 fi
 
-if [ -f ${AEROSPIKE_INCLUDE}/aerospike.h ]; then
-  printf "   [✓] ${AEROSPIKE_INCLUDE}/aerospike.h\n" >&1
+if [ -f ${AEROSPIKE_INCLUDE}/aerospike/aerospike.h ]; then
+  printf "   [✓] %s\n" "${AEROSPIKE_INCLUDE}/aerospike/aerospike.h" >&1
 else
-  printf "   [✗] ${AEROSPIKE_INCLUDE}/aerospike.h\n" >&1
+  printf "   [✗] %s\n" "${AEROSPIKE_INCLUDE}/aerospike/aerospike.h" >&1
   FAILED=1
 fi
 
 if [ -f ${AEROSPIKE_LUA}/aerospike.lua ]; then
-  printf "   [✓] ${AEROSPIKE_LUA}/aerospike.lua\n" >&1
+  printf "   [✓] %s\n" "${AEROSPIKE_LUA}/aerospike.lua" >&1
 else
-  printf "   [✗] ${AEROSPIKE_LUA}/aerospike.lua\n" >&1
+  printf "   [✗] %s\n" "${AEROSPIKE_LUA}/aerospike.lua" >&1
   FAILED=1
 fi
 
@@ -387,14 +265,14 @@ if [ $FAILED ]; then
   exit 1
 fi
 
-rm -rf ${AEROSPIKE}/lib
-mkdir -p ${AEROSPIKE}/lib
-cp ${PREFIX}/lib/libaerospike.a ${AEROSPIKE}/lib/.
+################################################################################
+# COPY FILES TO AEROSPIKE-C-CLIENT DIR
+################################################################################
 
-rm -rf ${AEROSPIKE}/include
-mkdir -p ${AEROSPIKE}/include
-cp -R ${PREFIX}/include/* ${AEROSPIKE}/include
-
-rm -rf ${AEROSPIKE}/lua
-mkdir -p ${AEROSPIKE}/lua
-cp -R ${AEROSPIKE_LUA}/* ${AEROSPIKE}/lua
+if [ $COPY_FILES == 1 ]; then
+  rm -rf ${AEROSPIKE}/{lib,include,lua}
+  mkdir -p ${AEROSPIKE}/{lib,include,lua}
+  cp ${AEROSPIKE_LIBRARY} ${AEROSPIKE}/lib/
+  cp -R ${AEROSPIKE_INCLUDE}/{aerospike,citrusleaf} ${AEROSPIKE}/include/
+  cp ${AEROSPIKE_LUA}/*.lua ${AEROSPIKE}/lua/
+fi